update submission result
